Changed paginator button creation
Paginator buttons can now contain anything whereas the previous implementation required the menu to have the form of the search menu. This means that I no longer have to make an exception in the main paginator class for items like the search submenu and that the class is far more useful for other menu types.