Fixed incorrect url encoding
Made sure that any content that the user can pass to the Overseerr api is properly encoded. This will cause searching to work a lot better than before.

Also fixed a bug where specifying the media type in the search command would cause the wrong type to be reported in the log and GUI.